+++ src/Factory/CompiledRouteRegistrarCacheFactory.php
@@ -4,16 +4,32 @@ declare(strict_types=1);
 
 namespace Sirix\Mezzio\Routing\Attributes\Factory;
 
+use Psr\Container\ContainerExceptionInterface;
+use Psr\Container\ContainerInterface;
+use Psr\Container\NotFoundExceptionInterface;
 use Sirix\Mezzio\Routing\Attributes\Cache\NullRouteRegistrarCache;
 use Sirix\Mezzio\Routing\Attributes\Cache\RouteCacheGenerator;
 use Sirix\Mezzio\Routing\Attributes\Cache\RouteCacheLoader;
 use Sirix\Mezzio\Routing\Attributes\Cache\RouteCacheStorage;
 use Sirix\Mezzio\Routing\Attributes\Cache\RouteRegistrarCacheInterface;
 use Sirix\Mezzio\Routing\Attributes\CompiledRouteRegistrarCache;
+use Sirix\Mezzio\Routing\Attributes\Config\RoutingAttributesConfig;
 
 final class CompiledRouteRegistrarCacheFactory
 {
-    public function __invoke(?string $cacheFile): RouteRegistrarCacheInterface
+    /**
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
+     */
+    public function __invoke(ContainerInterface $container): RouteRegistrarCacheInterface
+    {
+        $rootConfig = $container->has('config') ? $container->get('config') : [];
+        $config = RoutingAttributesConfig::fromRootConfig($rootConfig);
+
+        return $this->createFromCacheFile($config->cacheFile);
+    }
+
+    public function createFromCacheFile(?string $cacheFile): RouteRegistrarCacheInterface
     {
         if (null === $cacheFile) {
             return new NullRouteRegistrarCache();